Solidity might have both right shifts
The https://docs.soliditylang.org/ is contraticting itself on right
shifts. It mentions [1] that solidity have bitwise shifts. But the
actual descritpion of right shift [2] describes it as arithmetic one.

There are mentions of both right shifts in the solidity source
code [3].

I guess there is no harm in adding it.

[1] https://docs.soliditylang.org/en/v0.8.30/cheatsheet.html#order-of-precedence-of-operators
[2] https://docs.soliditylang.org/en/v0.8.30/types.html#shifts
[3] https://github.com/ethereum/solidity/blob/584cdcc5d994778ae141fd82e5e616bbed36d781/docs/grammar/SolidityParser.g4#L396